+++ Rebus.Tests/DataBus/SqlServer/SqlServerDataBusStorageTest.cs
@@ -1,8 +1,0 @@
-﻿using NUnit.Framework;
-using Rebus.Tests.Contracts.DataBus;
-
-namespace Rebus.Tests.DataBus.SqlServer
-{
-    [TestFixture]
-    public class SqlServerDataBusStorageTest : GeneralDataBusStorageTests<SqlServerDataBusStorageFactory> { }
-}