+++ Rebus.Tests/Integration/TestRetry.cs
@@ -12,13 +12,13 @@ using Rebus.Retry.Simple;
 using Rebus.Routing.TypeBased;
 using Rebus.Tests.Contracts;
 using Rebus.Tests.Contracts.Extensions;
-using Rebus.Tests.Extensions;
-using Rebus.Transport.Msmq;
+using Rebus.Transport.InMem;
+
 #pragma warning disable 1998
 
 namespace Rebus.Tests.Integration
 {
-    [TestFixture, Category(Categories.Msmq)]
+    [TestFixture]
     public class TestRetry : FixtureBase
     {
         static readonly string InputQueueName = TestConfig.GetName($"test.rebus2.retries.input@{Environment.MachineName}");
@@ -26,25 +26,22 @@ namespace Rebus.Tests.Integration
 
         BuiltinHandlerActivator _handlerActivator;
         IBus _bus;
+        InMemNetwork _network;
 
         void InitializeBus(int numberOfRetries)
         {
+            _network = new InMemNetwork();
+
             _handlerActivator = new BuiltinHandlerActivator();
 
             _bus = Configure.With(_handlerActivator)
                 .Logging(l => l.Console(minLevel: LogLevel.Warn))
-                .Transport(t => t.UseMsmq(InputQueueName))
+                .Transport(t => t.UseInMemoryTransport(_network, InputQueueName))
                 .Routing(r => r.TypeBased().Map<string>(InputQueueName))
                 .Options(o => o.SimpleRetryStrategy(maxDeliveryAttempts: numberOfRetries, errorQueueAddress: ErrorQueueName))
                 .Start();
 
             Using(_bus);
-        }
-
-        protected override void TearDown()
-        {
-            MsmqUtil.Delete(InputQueueName);
-            MsmqUtil.Delete(ErrorQueueName);
         }
 
         [Test]
@@ -64,14 +61,11 @@ namespace Rebus.Tests.Integration
 
             await _bus.Send("hej");
 
-            using (var errorQueue = new MsmqTransport(ErrorQueueName, new ConsoleLoggerFactory(true)))
-            {
-                var failedMessage = await errorQueue.AwaitReceive();
+            var failedMessage = await _network.WaitForNextMessageFrom(ErrorQueueName);
 
-                Assert.That(attemptedDeliveries, Is.EqualTo(numberOfRetries));
-                Assert.That(failedMessage.Headers.GetValue(Headers.ErrorDetails), Contains.Substring("5 unhandled exceptions"));
-                Assert.That(failedMessage.Headers.GetValue(Headers.SourceQueue), Is.EqualTo(InputQueueName));
-            }
+            Assert.That(attemptedDeliveries, Is.EqualTo(numberOfRetries));
+            Assert.That(failedMessage.Headers.GetValue(Headers.ErrorDetails), Contains.Substring("5 unhandled exceptions"));
+            Assert.That(failedMessage.Headers.GetValue(Headers.SourceQueue), Is.EqualTo(InputQueueName));
         }
 
         [TestCase(1)]
@@ -91,14 +85,11 @@ namespace Rebus.Tests.Integration
 
             await _bus.Send("hej");
 
-            using (var errorQueue = new MsmqTransport(ErrorQueueName, new ConsoleLoggerFactory(true)))
-            {
-                var expectedNumberOfAttemptedDeliveries = numberOfRetries;
+            await _network.WaitForNextMessageFrom(ErrorQueueName);
 
-                await errorQueue.AwaitReceive(2 + numberOfRetries / 10.0);
+            var expectedNumberOfAttemptedDeliveries = numberOfRetries;
 
-                Assert.That(attemptedDeliveries, Is.EqualTo(expectedNumberOfAttemptedDeliveries));
-            }
+            Assert.That(attemptedDeliveries, Is.EqualTo(expectedNumberOfAttemptedDeliveries));
         }
     }
 }