+++ Rebus.Tests/Transport/SqlServer/SqlServerTestManyMessages.cs
@@ -1,9 +1,0 @@
-﻿using NUnit.Framework;
-using Rebus.Tests.Contracts.Transports;
-using Rebus.Tests.Integration.ManyMessages;
-
-namespace Rebus.Tests.Transport.SqlServer
-{
-    [TestFixture]
-    public class SqlServerTestManyMessages : TestManyMessages<SqlServerBusFactory> { }
-}