+++ Rebus/Legacy/LegacyFlag.cs
@@ -1,0 +1,7 @@
+﻿namespace Rebus.Legacy
+{
+    /// <summary>
+    /// Flag that is registered in the Injectionist when legacy mode is active
+    /// </summary>
+    public class LegacyFlag { }
+}